fix: bet processing
Added own RPC, Added initial transaction processing upon railgun initialization
